# MM-27149: optimize initBasic