--- a/api4/bleve_test.go
+++ b/api4/bleve_test.go
@@ -10,7 +10,7 @@ import (
 )
 
 func TestBlevePurgeIndexes(t *testing.T) {
-	th := Setup(t).InitBasic()
+	th := Setup(t)
 	defer th.TearDown()
 
 	t.Run("as system user", func(t *testing.T) {