--- a/api4/config_test.go
+++ b/api4/config_test.go
@@ -15,7 +15,7 @@ import (
 )
 
 func TestGetConfig(t *testing.T) {
-	th := Setup(t).InitBasic()
+	th := Setup(t)
 	defer th.TearDown()
 	Client := th.Client
 
@@ -54,7 +54,7 @@ func TestGetConfig(t *testing.T) {
 }
 
 func TestReloadConfig(t *testing.T) {
-	th := Setup(t).InitBasic()
+	th := Setup(t)
 	defer th.TearDown()
 	Client := th.Client
 
@@ -80,7 +80,7 @@ func TestReloadConfig(t *testing.T) {
 }
 
 func TestUpdateConfig(t *testing.T) {
-	th := Setup(t).InitBasic()
+	th := Setup(t)
 	defer th.TearDown()
 	Client := th.Client
 
@@ -179,7 +179,7 @@ func TestUpdateConfig(t *testing.T) {
 }
 
 func TestUpdateConfigMessageExportSpecialHandling(t *testing.T) {
-	th := Setup(t).InitBasic()
+	th := Setup(t)
 	defer th.TearDown()
 
 	messageExportEnabled := *th.App.Config().MessageExportSettings.EnableExport
@@ -247,7 +247,7 @@ func TestUpdateConfigMessageExportSpecialHandling(t *testing.T) {
 }
 
 func TestUpdateConfigRestrictSystemAdmin(t *testing.T) {
-	th := Setup(t).InitBasic()
+	th := Setup(t)
 	defer th.TearDown()
 	th.App.UpdateConfig(func(cfg *model.Config) { *cfg.ExperimentalSettings.RestrictSystemAdmin = true })
 
@@ -293,7 +293,7 @@ func TestGetEnvironmentConfig(t *testing.T) {
 	defer os.Unsetenv("MM_SERVICESETTINGS_SITEURL")
 	defer os.Unsetenv("MM_SERVICESETTINGS_ENABLECUSTOMEMOJI")
 
-	th := Setup(t).InitBasic()
+	th := Setup(t)
 	defer th.TearDown()
 
 	t.Run("as system admin", func(t *testing.T) {
@@ -350,7 +350,7 @@ func TestGetEnvironmentConfig(t *testing.T) {
 }
 
 func TestGetOldClientConfig(t *testing.T) {
-	th := Setup(t).InitBasic()
+	th := Setup(t)
 	defer th.TearDown()
 
 	testKey := "supersecretkey"
@@ -402,7 +402,7 @@ func TestGetOldClientConfig(t *testing.T) {
 }
 
 func TestPatchConfig(t *testing.T) {
-	th := Setup(t).InitBasic()
+	th := Setup(t)
 	defer th.TearDown()
 
 	t.Run("config is missing", func(t *testing.T) {