--- a/api4/job_test.go
+++ b/api4/job_test.go
@@ -12,7 +12,7 @@ import (
 )
 
 func TestCreateJob(t *testing.T) {
-	th := Setup(t).InitBasic()
+	th := Setup(t)
 	defer th.TearDown()
 
 	job := &model.Job{
@@ -39,7 +39,7 @@ func TestCreateJob(t *testing.T) {
 }
 
 func TestGetJob(t *testing.T) {
-	th := Setup(t).InitBasic()
+	th := Setup(t)
 	defer th.TearDown()
 
 	job := &model.Job{
@@ -68,7 +68,7 @@ func TestGetJob(t *testing.T) {
 }
 
 func TestGetJobs(t *testing.T) {
-	th := Setup(t).InitBasic()
+	th := Setup(t)
 	defer th.TearDown()
 
 	jobType := model.NewId()
@@ -115,7 +115,7 @@ func TestGetJobs(t *testing.T) {
 }
 
 func TestGetJobsByType(t *testing.T) {
-	th := Setup(t).InitBasic()
+	th := Setup(t)
 	defer th.TearDown()
 
 	jobType := model.NewId()
@@ -173,7 +173,7 @@ func TestGetJobsByType(t *testing.T) {
 }
 
 func TestCancelJob(t *testing.T) {
-	th := Setup(t).InitBasic()
+	th := Setup(t)
 	defer th.TearDown()
 
 	jobs := []*model.Job{