--- a/api4/post_test.go
+++ b/api4/post_test.go
@@ -1004,7 +1004,7 @@ func TestPinPost(t *testing.T) {
 }
 
 func TestUnpinPost(t *testing.T) {
-	th := Setup(t)
+	th := Setup(t).InitBasic()
 	defer th.TearDown()
 	Client := th.Client