--- a/api4/saml_test.go
+++ b/api4/saml_test.go
@@ -13,7 +13,7 @@ import (
 )
 
 func TestGetSamlMetadata(t *testing.T) {
-	th := Setup(t).InitBasic()
+	th := Setup(t)
 	defer th.TearDown()
 	Client := th.Client