@agnivade 2020-07-21
@lindy65 - This just improves the test duration. So no need to test this one.